+++ examples/src/JShark/Example/Hvm2Demo.hs
@@ -1,4 +1,0 @@
-module JShark.Example.Hvm2Demo (page, mainJS) where
-
-import JShark.Example.Hvm2Demo.Client (mainJS)
-import JShark.Example.Hvm2Demo.Page (page)