+++ wasm/hvm2/include/hvm2_wasm.h
@@ -1,6 +1,0 @@
-/* Optional WASM shims when targeting freestanding; unused on wasm32-wasi. */
-#pragma once
-
-#ifndef PRIu64
-#define PRIu64 "llu"
-#endif